refactor(edge): isolate load balancing key resolution helpers
